% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state represents geographically the values in a given year and a given state number
which have a longitude > 900 and a latitude > 90.
It will thrown an error if the year or the state are not valid, or if there are no accidents to show.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{the number of the selected state}

\item{year}{the selected year}
}
\value{
The function returns a graphic showing the geographical representation of the data
 for a given year and a given state.
}
\description{
fars_map_state represents geographically the values in a given year and a given state number
which have a longitude > 900 and a latitude > 90.
It will thrown an error if the year or the state are not valid, or if there are no accidents to show.
}
\note{
The original function has been adapted to Non-standard evaluation.
For example, the filter function below has been modified to filter_
}
\examples{
\dontrun{fars_map_state(1, 2013)}
}
