% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This function, named "fars_read", reads a file and save the data in a variable called "data".}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A single string with the name of the file.
Only years 2013, 2014 and 2015 are available. Trying loading a different year will result in
a message saying the file does not exist.}
}
\value{
The function returns a data frame containing the data of the file. The output
file is named "data".
}
\description{
This function, named "fars_read", reads a file and save the data in a variable called "data".
}
\examples{
\dontrun{dirdata <- system.file("extdata", "accident_2013.csv.bz2", package = "FARSinR2")}
\dontrun{fars_read(dirdata)}
}
