% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years function takes a series of years and create an ordenated table for a designated month.
The core of the function uses tryCatch, a function for
signaling and handling unusual conditions, including errors and warnings.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector containing the selected years.}
}
\value{
The function returns a data frame containing the month and year columns of the file.
It will thrown an error if the year is not valid; the only valid years using the provided data set are 2013, 2014 and 2015.
c(2013,2014) is OK, whereas c(2016,2017) will show NULL and a warning message.
}
\description{
fars_read_years function takes a series of years and create an ordenated table for a designated month.
The core of the function uses tryCatch, a function for
signaling and handling unusual conditions, including errors and warnings.
}
\examples{
\dontrun{fars_read_years(c(2013,2014))}
}
