% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years sumarizes in a monthly basis the accidents in a given years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{a vector containing the selected years, for example c(2013, 2014)}
}
\value{
The function returns a data frame containing the month and the values for different years in their respective columns.
For example:
  MONTH `2013` `2014`
  *  <int>  <int>  <int>
  1      1   2230   2168
  2      2   1952   1893
}
\description{
fars_summarize_years sumarizes in a monthly basis the accidents in a given years.
}
\examples{
\dontrun{fars_summarize_years(c(2013,2014))}
}
