% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARSinR2.R
\docType{package}
\name{FARSinR2-package}
\alias{FARSinR2}
\alias{FARSinR2-package}
\title{FARSinR2 - second try of the package.}
\description{
FARSinR2 - second try of the package.
}
\author{
\strong{Maintainer}: Alfonso Delgado-Bonal \email{alfonso.delgado.bonal@gmail.com}

}
