% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{The make_filename function returns a character vector for the file name, taking the year as argumnent.
All values are accepted, no error will be thrown.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a value which will be converted to integer.}
}
\value{
the output is a character vector which includes the argument, customizing the file name.
}
\description{
The make_filename function returns a character vector for the file name, taking the year as argumnent.
All values are accepted, no error will be thrown.
}
\examples{
\dontrun{make_filename(2014)}

}
